requirements.txt
wheel
django
djangorestframework
pytils
pillow
drf-yasg
model_bakery
django_filter
django-cors-headers